Move extensions to simple local functions
